chore: make modern-compiler tests as default
